Move IntlProvider back in [lang] dir
